requirements.txt
kafka-python==2.3.0
pandas==3.0.1
pyspark==3.5.0
google-cloud-bigquery==3.40.1
google-cloud-storage==3.10.1
streamlit
plotly
dbt-bigquery
pyarrow